+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/infrastructure/api/bff/PriorNotificationController.kt
@@ -26,7 +26,6 @@ class PriorNotificationController(
     private val getPriorNotificationTypes: GetPriorNotificationTypes,
     private val updatePriorNotificationNote: UpdatePriorNotificationNote,
     private val verifyAndSendPriorNotification: VerifyAndSendPriorNotification,
-    private val getPriorNotificationPdfDocument: GetPriorNotificationPdfDocument,
 ) {
     @GetMapping("")
     @Operation(summary = "Get all prior notifications")
