+++ backend/src/test/kotlin/fr/gouv/cnsp/monitorfish/infrastructure/api/bff/PriorNotificationControllerITests.kt
@@ -6,8 +6,6 @@ import com.nhaarman.mockitokotlin2.anyOrNull
 import fr.gouv.cnsp.monitorfish.config.SentryConfig
 import fr.gouv.cnsp.monitorfish.domain.entities.logbook.LogbookMessagePurpose
 import fr.gouv.cnsp.monitorfish.domain.entities.prior_notification.ManualPriorNotificationComputedValues
-import fr.gouv.cnsp.monitorfish.domain.entities.prior_notification.PdfDocument
-import fr.gouv.cnsp.monitorfish.domain.entities.prior_notification.PriorNotificationSource
 import fr.gouv.cnsp.monitorfish.domain.entities.prior_notification.PriorNotificationStats
 import fr.gouv.cnsp.monitorfish.domain.use_cases.prior_notification.*
 import fr.gouv.cnsp.monitorfish.domain.utils.PaginatedList
@@ -26,13 +24,14 @@ import org.springframework.context.annotation.Import
 import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.*
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import java.time.ZonedDateTime
 
 @Import(SentryConfig::class)
 @AutoConfigureMockMvc(addFilters = false)
 @WebMvcTest(value = [(PriorNotificationController::class)])
-class PublicPriorNotificationControllerITests {
+class PriorNotificationControllerITests {
     @Autowired
     private lateinit var api: MockMvc
 
@@ -56,9 +55,6 @@ class PublicPriorNotificationControllerITests {
 
     @MockBean
     private lateinit var updatePriorNotificationNote: UpdatePriorNotificationNote
-
-    @MockBean
-    private lateinit var getPriorNotificationPdfDocument: GetPriorNotificationPdfDocument
 
     @Autowired
     private lateinit var objectMapper: ObjectMapper
@@ -339,61 +335,5 @@ class PublicPriorNotificationControllerITests {
                     equalTo(fakePriorNotification.logbookMessageTyped.typedMessage.note),
                 ),
             )
-    }
-
-    @Test
-    fun `getPdf Should get the PDF of a prior notification`() {
-        val dummyPdfContent = """
-            %PDF-1.4
-            1 0 obj
-            << /Type /Catalog /Pages 2 0 R >>
-            endobj
-            2 0 obj
-            << /Type /Pages /Kids [3 0 R] /Count 1 >>
-            endobj
-            3 0 obj
-            << /Type /Page /Parent 2 0 R /MediaBox [0 0 612 792] /Contents 4 0 R >>
-            endobj
-            4 0 obj
-            << /Length 55 >>
-            stream
-            BT
-            /F1 24 Tf
-            100 700 Td
-            (Hello, World!) Tj
-            ET
-            endstream
-            endobj
-            xref
-            0 5
-            0000000000 65535 f
-            0000000010 00000 n
-            0000000062 00000 n
-            0000000113 00000 n
-            0000000218 00000 n
-            trailer
-            << /Size 5 /Root 1 0 R >>
-            startxref
-            291
-            %%EOF
-        """.trimIndent().toByteArray()
-
-        // Given
-        given(getPriorNotificationPdfDocument.execute("REPORT_ID"))
-            .willReturn(
-                PdfDocument(
-                    reportId = "REPORT_ID",
-                    source = PriorNotificationSource.LOGBOOK,
-                    generationDatetimeUtc = ZonedDateTime.now(),
-                    pdfDocument = dummyPdfContent,
-                ),
-            )
-
-        // When
-        api.perform(get("/api/v1/prior_notifications/pdf/REPORT_ID"))
-            // Then
-            .andExpect(status().isOk)
-            .andExpect(content().contentType(MediaType.APPLICATION_PDF))
-            .andExpect(content().bytes(dummyPdfContent))
     }
 }
